refactor: Move auth-services to ~/context
