+++ packages/server/src/internals/authentication/index.ts
@@ -1,2 +1,1 @@
-export * from './auth-services'
 export * from './handle-authentication'
