+++ packages/server/src/internals/server/index.ts
@@ -4,10 +4,10 @@ import { Pool, createPool } from 'mysql2/promise'
 
 import { applyGraphQLMiddleware } from './graphql-middleware'
 import { applySwrQueueDashboardMiddleware } from './swr-queue-dashboard-middleware'
-import { createAuthServices, AuthServices } from '../authentication'
 import { createCache, createEmptyCache } from '../cache'
 import { initializeSentry } from '../sentry'
 import { createSwrQueue } from '../swr-queue'
+import { createAuthServices, AuthServices } from '~/context/auth-services'
 import { Cache } from '~/context/cache'
 import { SwrQueue } from '~/context/swr-queue'
 import { applyEnmeshedMiddleware } from '~/internals/server/enmeshed-middleware'
