+++ packages/server/src/internals/server/kratos-middleware.ts
@@ -5,7 +5,7 @@ import * as t from 'io-ts'
 import { JwtPayload, decode } from 'jsonwebtoken'
 import { validate as uuidValidate } from 'uuid'
 
-import { Kratos } from '../authentication'
+import { Kratos } from '~/context/auth-services'
 import { captureErrorEvent } from '~/error-event'
 import { createRequest } from '~/internals/data-source-helper'
 import { DatabaseLayer } from '~/model'
