+++ packages/server/src/internals/swr-queue.ts
@@ -3,10 +3,10 @@ import { either as E, option as O } from 'fp-ts'
 import * as t from 'io-ts'
 import * as R from 'ramda'
 
-import { createAuthServices } from './authentication'
 import { isLegacyQuery, LegacyQuery } from './data-source-helper'
 import { log } from './log'
 import { type Context } from '~/context'
+import { createAuthServices } from '~/context/auth-services'
 import { CacheEntry, Cache, Priority } from '~/context/cache'
 import { SwrQueue } from '~/context/swr-queue'
 import { Database } from '~/database'
